feat: Implement adding yearly repetition

+++ src/main/java/ru/steqa/repetitionservice/scheme/rabbit/repetition/FixedYearRepetition.java
@@ -7,11 +7,11 @@ import ru.steqa.repetitionservice.scheme.rabbit.TransactionType;
 import java.time.LocalDateTime;
 
 public class FixedYearRepetition extends BaseRepetitionRule {
-    public Byte day;
-    public Byte month;
+    public Integer day;
+    public Integer month;
 
     public FixedYearRepetition(Long userId, Long transactionId, TransactionType transactionType,
-                               LocalDateTime nextExecution, Byte day, Byte month) {
+                               LocalDateTime nextExecution, Integer day, Integer month) {
         super(userId, transactionId, transactionType, nextExecution, RepetitionMode.FIXED_YEAR);
         this.day = day;
         this.month = month;
